Move tree view functionality into common file
It is possible to avoid the import loop by using the TYPE_CHECKING
constant from the typing module to avoid the import of TreeViewSheet at
runtime, because it is only used for the type comment.
