+++ plugin/core/registry.py
@@ -3,8 +3,6 @@ from .protocol import Location
 from .protocol import LocationLink
 from .sessions import AbstractViewListener
 from .sessions import Session
-from .tree_view import TreeDataProvider
-from .tree_view import TreeViewSheet
 from .typing import Optional, Any, Generator, Iterable, List, Union
 from .views import first_selection_region
 from .views import get_uri_and_position_from_location
@@ -16,54 +14,10 @@ from .windows import WindowRegistry
 from functools import partial
 import operator
 import sublime
-import sublime_api  # pyright: ignore[reportMissingImports]
 import sublime_plugin
 
 
 windows = WindowRegistry()
-
-
-def new_tree_view_sheet(
-    window: sublime.Window,
-    name: str,
-    data_provider: TreeDataProvider,
-    header: str = "",
-    flags: int = 0,
-    group: int = -1
-) -> Optional[TreeViewSheet]:
-    """
-    Use this function to create a new TreeView in form of a special HtmlSheet (TreeViewSheet). Only one TreeViewSheet
-    with the given name is allowed per window. If there already exists a TreeViewSheet with the same name, its content
-    will be replaced with the new data. The header argument is allowed to contain minihtml markup.
-    """
-    wm = windows.lookup(window)
-    if not wm:
-        return None
-    if name in wm.tree_view_sheets:
-        tree_view_sheet = wm.tree_view_sheets[name]
-        sheet_id = tree_view_sheet.id()
-        if tree_view_sheet.window():
-            tree_view_sheet.set_provider(data_provider, header)
-            if flags & sublime.ADD_TO_SELECTION:
-                # add to selected sheets if not already selected
-                selected_sheets = window.selected_sheets()
-                for sheet in window.sheets():
-                    if isinstance(sheet, sublime.HtmlSheet) and sheet.id() == sheet_id:
-                        if sheet not in selected_sheets:
-                            selected_sheets.append(sheet)
-                            window.select_sheets(selected_sheets)
-                        break
-            else:
-                window.focus_sheet(tree_view_sheet)
-            return tree_view_sheet
-    tree_view_sheet = TreeViewSheet(
-        sublime_api.window_new_html_sheet(window.window_id, name, "", flags, group),
-        name,
-        data_provider,
-        header
-    )
-    wm.tree_view_sheets[name] = tree_view_sheet
-    return tree_view_sheet
 
 
 def best_session(view: sublime.View, sessions: Iterable[Session], point: Optional[int] = None) -> Optional[Session]:
@@ -320,28 +274,3 @@ class LspPrevDiagnosticCommand(LspTextCommand):
 
     def run(self, edit: sublime.Edit, point: Optional[int] = None) -> None:
         navigate_diagnostics(self.view, point, forward=False)
-
-
-def toggle_tree_item(window: sublime.Window, name: str, id: str, expand: bool) -> None:
-    wm = windows.lookup(window)
-    if not wm:
-        return
-    sheet = wm.tree_view_sheets.get(name)
-    if not sheet:
-        return
-    if expand:
-        sheet.expand_item(id)
-    else:
-        sheet.collapse_item(id)
-
-
-class LspExpandTreeItemCommand(LspWindowCommand):
-
-    def run(self, name: str, id: str) -> None:
-        toggle_tree_item(self.window, name, id, True)
-
-
-class LspCollapseTreeItemCommand(LspWindowCommand):
-
-    def run(self, name: str, id: str) -> None:
-        toggle_tree_item(self.window, name, id, False)
