+++ plugin/hierarchy.py
@@ -13,8 +13,8 @@ from .core.protocol import TypeHierarchyPrepareParams
 from .core.registry import get_position
 from .core.registry import LspTextCommand
 from .core.registry import LspWindowCommand
-from .core.registry import new_tree_view_sheet
 from .core.sessions import Session
+from .core.tree_view import new_tree_view_sheet
 from .core.tree_view import TreeDataProvider
 from .core.tree_view import TreeItem
 from .core.typing import Callable, List, Optional, TypedDict, Union
